update: test match sorting
